requirements.txt
tqdm==4.31.1
lxml
requests>=2.20.0
psutil>=5.6.6
subprocess32==3.5.4
wget==3.2
biopython
mysolr
